DocumentProcessing/document-processing-pipeline/simple-pdf-extraction/requirements.txt
# PDF processing
pypdf2==3.0.1
PyMuPDF==1.23.18
pdfplumber==0.10.3

# Document and text processing
python-docx==1.1.0
python-pptx==0.6.23

# OCR support (for scanned PDFs)
pytesseract==0.3.10
Pillow==10.2.0

# HTTP client for Ollama
requests==2.31.0

# Logging and utilities
python-dotenv==1.0.0
tqdm==4.66.1

# File handling (mimetypes is built-in, no need to install)